+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-06-24 14:47:12
-// Commit: test(i18n): implement widget integration tests for i18n
+// Updated at: 2026-06-24 21:52:54
+// Commit: style(i18n): harmonize light and dark theme palette for i18n
 
 class i18nComponentState {
-  final String commitRef = '2259';
-  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
-  final DateTime timestamp = DateTime.parse('2026-06-24T14:47:12+05:00');
+  final String commitRef = '2281';
+  final String lastMessage = 'style(i18n): harmonize light and dark theme palette for i18n';
+  final DateTime timestamp = DateTime.parse('2026-06-24T21:52:54+05:00');
 }